Move Remote and Receiver to a new remote_stream module
Make Remote and Receiver generic over the response future and body,
instead of the http service.
